Researching performance issues between fixing-interpreter and dev

+++ src/main/java/envision_lang/interpreter/util/CastingUtil.java
@@ -174,7 +174,7 @@ public class CastingUtil {
 			return switch (to) {
 			//case CHAR -> EnvisionCharClass.newChar(doubleVal);
 			//case BOOLEAN -> EnvisionBooleanClass.newBoolean(doubleVal);
-			case INT -> EnvisionIntClass.newInt(doubleVal);
+			case INT -> EnvisionIntClass.newInt((long) doubleVal);
 			case NUMBER, DOUBLE -> EnvisionDoubleClass.newDouble(doubleVal);
 			case STRING -> EnvisionStringClass.newString(doubleVal);
 			default -> throw new VariableCastError("Invalid cast type: " + typeIn + "!");
